"Made the home page the main activity and made the today page it's own different activity. The main page opens the today, symptoms and health pages if you press their respective buttons"
