+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
+
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -29,14 +30,17 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name=".Today"
+            android:exported="true"></activity>
 
-        <receiver android:name=".NotificationAlarmManager"/>
+        <receiver android:name=".NotificationAlarmManager" />
 
         <activity
             android:name=".SymptomsInputActivity"
-            android:screenOrientation="portrait">
-        </activity>
-
+            android:screenOrientation="portrait"></activity>
     </application>
+
+
 
 </manifest>
