+++ app/src/main/java/com/example/periodtracker/MainActivity.java
@@ -1,134 +1,53 @@
 package com.example.periodtracker;
 
-
-import android.Manifest;
-import android.os.Bundle;
-import androidx.activity.result.ActivityResultLauncher;
-import androidx.activity.result.contract.ActivityResultContracts;
-import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
+import android.os.Bundle;
 import android.view.View;
-import android.view.animation.AlphaAnimation;
 import android.widget.Button;
-import android.widget.ImageView;
 import android.widget.TextView;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.datepicker.MaterialDatePicker;
-import com.google.android.material.datepicker.MaterialPickerOnPositiveButtonClickListener;
-import java.time.Instant;
+
+import androidx.appcompat.app.AppCompatActivity;
+
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
-import java.util.Calendar;
 import java.util.Locale;
 
 public class MainActivity extends AppCompatActivity {
-
-    private TextView displayDay;
-    private TextView displayMonth;
-    private Calendar currentDate;
-    private Locale locale;
-    private AlphaAnimation buttonClickAnimation;
-
-    /**
-     * Required permissions for the app to work.
-     */
-    static final String[] REQUIRED_PERMISSIONS = {
-            Manifest.permission.POST_NOTIFICATIONS,
-    };
-
-    private TextView minPrediction;
-    private TextView maxPrediction;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        minPrediction = findViewById(R.id.minPredictionText);
-        maxPrediction = findViewById(R.id.maxPredictionText);
-
-        buttonClickAnimation = new AlphaAnimation(1,0);
-        buttonClickAnimation.setDuration(300);
-        currentDate = Calendar.getInstance();
-        locale = Locale.getDefault();
-        displayDay = findViewById(R.id.day);
-        displayMonth = findViewById(R.id.month);
-        displayDay.setText(String.valueOf(currentDate.get(Calendar.DAY_OF_MONTH)));
-        displayMonth.setText(currentDate.getDisplayName(Calendar.MONTH, Calendar.SHORT, locale));
-
-        Button btn_tick = findViewById(R.id.btn_tick);
-        btn_tick.setOnClickListener(v -> {
-            ImageView periodBtnTick = findViewById(R.id.periodBtnTick);
-            v.startAnimation(buttonClickAnimation);
-            int visibility = periodBtnTick.getVisibility();
-            if(visibility == View.VISIBLE)
-            {
-                periodBtnTick.startAnimation(buttonClickAnimation);
-                periodBtnTick.setVisibility(View.GONE);
-            }
-            else
-            {
-                periodBtnTick.setVisibility(View.VISIBLE);
-            }
-        });
-
-
-        // add this to the bottom of pages for navigation
-        // Initialize and assign variable
-        BottomNavigationView bottomNavigationView=findViewById(R.id.bottom_navigation);
-
-        // Set to highlight current page, edit item id according to button id
-        bottomNavigationView.setSelectedItemId(R.id.today);
-
-        // Perform item selected listener
-        bottomNavigationView.setOnItemSelectedListener(item -> {
-            if(item.getItemId() == R.id.health){
-                    startActivity(new Intent(getApplicationContext(),Health.class));
-                    overridePendingTransition(0,0);
-                return true;
-            }
-            else if(item.getItemId() == R.id.today){
-                return true;
-            }
-            else if(item.getItemId() == R.id.symptoms){
-                    startActivity(new Intent(getApplicationContext(), SymptomsActivity.class));
-                    overridePendingTransition(0,0);
-                return true;
-            }
-            else{
-                return false;
-            }
-        });
-
-        ActivityResultContracts.RequestMultiplePermissions resultContract = new ActivityResultContracts.RequestMultiplePermissions();
-        ActivityResultLauncher<String[]> multiplePermLauncher = registerForActivityResult(resultContract,
-                isGranted -> {
-                    // Any permission specific code put here:
-                });
-        multiplePermLauncher.launch(REQUIRED_PERMISSIONS);
-
+        LocalDate today = LocalDate.now();
+        LocalDate yest = today.minusDays(1);
+        LocalDate tomo = today.plusDays(1);
+        int month = today.getMonthValue();
+        final Button todayButton = findViewById(R.id.button);
+        final Button yestButton = findViewById(R.id.button2);
+        final Button tomoButton = findViewById(R.id.button3);
+        todayButton.setText(String.format(Integer.toString(today.getDayOfMonth())));
+        yestButton.setText(String.format(Integer.toString(yest.getDayOfMonth())));
+        tomoButton.setText(String.format(Integer.toString(tomo.getDayOfMonth())));
+        final TextView monthView = findViewById(R.id.textView);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMM", Locale.ENGLISH);
+        monthView.setText(today.format(formatter));
     }
 
-    public void onSelectDateClick(View view) {
-        MaterialDatePicker<Long> materialDatePicker = MaterialDatePicker.Builder.datePicker()
-                .setTitleText(getString(R.string.select_date))
-                .setSelection(MaterialDatePicker.todayInUtcMilliseconds())
-                .build();
-        materialDatePicker.addOnPositiveButtonClickListener(new MaterialPickerOnPositiveButtonClickListener<Long>() {
-            @Override
-            public void onPositiveButtonClick(Long selection) {
-                LocalDateTime dateTime = LocalDateTime.ofInstant(Instant.ofEpochMilli(selection), ZoneId.systemDefault());
-                LocalDate selectedDate = dateTime.toLocalDate();
-                MenstrualCycleTracker.PeriodPrediction prediction = MenstrualCycleTracker.trackMenstrualCycle(selectedDate);
-                DateTimeFormatter format = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-                minPrediction.setText(prediction.minEndDate.format(format));
-                maxPrediction.setText(prediction.maxEndDate.format(format));
-            }
-        });
+    public void onYestPressed(View v){}
 
-        materialDatePicker.show(getSupportFragmentManager(), "tag");
+    public void onTodayPressed(View v){}
+
+    public void onTomoPressed(View v){}
+
+    public void onEditPressed(View v){
+        startActivity(new Intent(this, Today.class));
     }
 
+    public void onHealthPressed(View v){
+        startActivity(new Intent(this, Health.class));
+    }
+
+    public void onSymptomsPressed(View v){
+        startActivity(new Intent(this, SymptomsActivity.class));
+    }
 }
